+++ magda-web-client/src/Components/Dataset/Add/DatasetAddMetadataPage.tsx
@@ -105,6 +105,7 @@ class NewDataset extends React.Component<Props, State> {
             <DatasetAddPeoplePage
                 edit={this.edit}
                 dataset={this.state.dataset}
+                publishing={this.state.datasetPublishing}
             />
         ),
         this.renderRestriction.bind(this),